Added comments in the code and restructured the code to make it simple
