+++ javaprocessor/src/main/java/simplewallet/application/SimpleWalletRequestProcessor.java
@@ -29,14 +29,16 @@ public class SimpleWalletRequestProcessor {
     }
 
     public boolean execute(TpProcessRequest transactionRequest, State stateInfo) throws InvalidTransactionException, InternalError {
-	//Get request payload data
+	// Get request payload data
 	String payload =  transactionRequest.getPayload().toStringUtf8();
 	ArrayList<String> payloadList = new ArrayList<>(Arrays.asList(payload.split(",")));
 	if(payloadList.size() != 2) {
 	    throw new InvalidTransactionException("Invalid no. of arguments: expected 2, got:" + payloadList.size());
 	}
+	// We would get the operation name and amount in the payload
 	String operation = payloadList.get(0);
 	this.amount = Integer.valueOf(payloadList.get(1));
+	// Getting the user signing piblic key from header
 	this.userKey = transactionRequest.getHeader().getSignerPublicKey();
 	boolean status = false;
 	switch(operation) {
@@ -55,16 +57,13 @@ public class SimpleWalletRequestProcessor {
     }
 
     private boolean makeDeposit(State stateInfo) throws InvalidTransactionException, InternalError {
+	//Get the wallet key from the signer public key
 	String walletKey = this.getWalletKey(this.userKey);
 	logger.info("Got user key" + this.userKey + "wallet key " + walletKey);
+	//Get balance from ledger state
 	String balance = stateInfo.getState(Collections.singletonList(walletKey)).get(walletKey).toStringUtf8();
 	Integer newBalance = 0;
-	// Update balance
-	if (this.amount <= 0) {
-	    String error = "Deposit amount should be greater than 0";
-	    throw new InvalidTransactionException(error);
-	}
-	else if (balance.isEmpty()) {
+	if (balance.isEmpty()) {
 	    logger.info("This is the first time we got a deposit for user.");
 	    logger.info("Creating a new account for the user: " + this.userKey);
 	    newBalance = this.amount;
@@ -72,6 +71,7 @@ public class SimpleWalletRequestProcessor {
 	else {
 	    newBalance = Integer.valueOf(balance) + this.amount;
 	}
+	// Update balance
 	Map.Entry<String, ByteString> entry = new AbstractMap.SimpleEntry<String, ByteString>(walletKey,
 		ByteString.copyFromUtf8(newBalance.toString()));
 	Collection<Map.Entry<String, ByteString>> ledgerEntry = Collections.singletonList(entry);
@@ -83,6 +83,7 @@ public class SimpleWalletRequestProcessor {
     private boolean makeWithdraw(State stateInfo) throws InvalidTransactionException, InternalError {
  	String walletKey = this.getWalletKey(this.userKey);
  	logger.info("Got user key " + this.userKey + "wallet key "+ walletKey);
+	//Get balance from ledger state
  	String balance = stateInfo.getState(Collections.singletonList(walletKey)).get(walletKey).toStringUtf8();
 
  	if (balance.isEmpty()) {
@@ -104,6 +105,7 @@ public class SimpleWalletRequestProcessor {
      }
 
     private String getWalletKey(String userKey) {
+	//Generate unique key(wallet key) from the wallet namespace and user signer key
 	return Utils.hash512(SimpleWalletNameSpace.getBytes()).substring(0, 6)
 		+ Utils.hash512(userKey.getBytes()).substring(0, 64);
     }
